Refactored code, added test hook, addressed PR comments
