+++ src/Microsoft.PowerShell.ConsoleHost/host/msh/Executor.cs
@@ -324,12 +324,16 @@ namespace Microsoft.PowerShell
 
             // Experimental:
             // Check for implicit remoting commands that can be batched, and execute as batched if able.
-            if (ExperimentalFeature.EnabledExperimentalFeatureNames.TryGetValue("PSImplicitRemotingBatching", out string actualValue))
+            if (ExperimentalFeature.IsEnabled("PSImplicitRemotingBatching"))
             {
                 var addOutputter = ((options & ExecutionOptions.AddOutputter) > 0);
-                exceptionThrown = null;
-                if (addOutputter && TryRunAsImplicitBatch(command, out exceptionThrown))
+                if (addOutputter &&
+                    !_parent.RunspaceRef.IsRunspaceOverridden &&
+                    _parent.RunspaceRef.Runspace.ExecutionContext.Modules != null &&
+                    _parent.RunspaceRef.Runspace.ExecutionContext.Modules.IsImplicitRemotingModuleLoaded &&
+                    Utils.TryRunAsImplicitBatch(command, _parent.RunspaceRef.Runspace))
                 {
+                    exceptionThrown = null;
                     return null;
                 }
             }
@@ -338,329 +342,6 @@ namespace Microsoft.PowerShell
 
             return ExecuteCommandHelper(tempPipeline, out exceptionThrown, options);
         }
-
-        #region ImplicitRemotingBatching
-
-        // A visitor to walk an AST and validate that it is a candidate for implicit remoting batching.
-        // Based on ScriptBlockToPowerShellChecker.
-        internal class PipelineForBatchingChecker : AstVisitor
-        {
-            internal readonly HashSet<string> ValidVariables = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-            internal readonly HashSet<string> Commands = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-            internal ScriptBlockAst ScriptBeingConverted { get; set; }
-
-            public override AstVisitAction VisitVariableExpression(VariableExpressionAst variableExpressionAst)
-            {
-                if (!variableExpressionAst.VariablePath.IsAnyLocal())
-                {
-                    ThrowError(
-                            new ImplicitRemotingBatchingNotSupportedException(
-                                ConsoleHostStrings.ImplicitBatchVariableTypeNotSupported),
-                            variableExpressionAst);
-                }
-
-                if (variableExpressionAst.VariablePath.UnqualifiedPath != "_")
-                {
-                    ValidVariables.Add(variableExpressionAst.VariablePath.UnqualifiedPath);
-                }
-
-                return AstVisitAction.Continue;
-            }
-
-            public override AstVisitAction VisitPipeline(PipelineAst pipelineAst)
-            {
-                if (pipelineAst.PipelineElements[0] is CommandExpressionAst)
-                {
-                    // If the first element is a CommandExpression, this pipeline should be the value
-                    // of a parameter. We want to avoid a scriptblock that contains only a pure expression.
-                    // The check "pipelineAst.Parent.Parent == ScriptBeingConverted" guarantees we throw
-                    // error on that kind of scriptblock.
-
-                    // Disallow pure expressions at the "top" level, but allow them otherwise.
-                    // We want to catch:
-                    //     1 | echo
-                    // But we don't want to error out on:
-                    //     echo $(1)
-                    // See the comment in VisitCommand on why it's safe to check Parent.Parent, we
-                    // know that we have at least:
-                    //     * a NamedBlockAst (the end block)
-                    //     * a ScriptBlockAst (the ast we're comparing to)
-                    if (pipelineAst.GetPureExpression() == null || pipelineAst.Parent.Parent == ScriptBeingConverted)
-                    {
-                        ThrowError(
-                            new ImplicitRemotingBatchingNotSupportedException(
-                                ConsoleHostStrings.ImplicitBatchStartExpressionNotSupported),
-                            pipelineAst);
-                    }
-                }
-
-                return AstVisitAction.Continue;
-            }
-
-            public override AstVisitAction VisitCommand(CommandAst commandAst)
-            {
-                if (commandAst.InvocationOperator == TokenKind.Dot)
-                {
-                    ThrowError(
-                        new ImplicitRemotingBatchingNotSupportedException(
-                            ConsoleHostStrings.ImplicitBatchDotSourcingNotSupported),
-                        commandAst);
-                }
-
-                /*
-                // Up front checking ensures that we have a simple script block,
-                // so we can safely assume that the parents are:
-                //     * a PipelineAst
-                //     * a NamedBlockAst (the end block)
-                //     * a ScriptBlockAst (the ast we're comparing to)
-                // If that isn't the case, the conversion isn't allowed.  It
-                // is also safe to assume that we have at least 3 parents, a script block can't be simpler.
-                if (commandAst.Parent.Parent.Parent != ScriptBeingConverted)
-                {
-                    ThrowError(
-                        new ImplicitRemotingBatchingNotSupportedException(
-                            "CantConvertWithCommandInvocations not supported"),
-                        commandAst);
-                }
-                */
-
-                if (commandAst.CommandElements[0] is ScriptBlockExpressionAst)
-                {
-                    ThrowError(
-                        new ImplicitRemotingBatchingNotSupportedException(
-                            ConsoleHostStrings.ImplicitBatchScriptBlockNotSupported),
-                        commandAst);
-                }
-
-                var commandName = commandAst.GetCommandName();
-                if (commandName != null)
-                {
-                    Commands.Add(commandName);
-                }
-
-                return AstVisitAction.Continue;
-            }
-
-            public override AstVisitAction VisitMergingRedirection(MergingRedirectionAst redirectionAst)
-            {
-                if (redirectionAst.ToStream != RedirectionStream.Output)
-                {
-                    ThrowError(
-                        new ImplicitRemotingBatchingNotSupportedException(
-                            ConsoleHostStrings.ImplicitBatchMergeRedirectNotSupported),
-                        redirectionAst);
-                }
-
-                return AstVisitAction.Continue;
-            }
-
-            public override AstVisitAction VisitFileRedirection(FileRedirectionAst redirectionAst)
-            {
-                ThrowError(
-                    new ImplicitRemotingBatchingNotSupportedException(
-                        ConsoleHostStrings.ImplicitBatchFileRedirectNotSupported),
-                    redirectionAst);
-
-                return AstVisitAction.Continue;
-            }
-
-            /*
-            public override AstVisitAction VisitScriptBlockExpression(ScriptBlockExpressionAst scriptBlockExpressionAst)
-            {
-                ThrowError(new ImplicitRemotingBatchingNotSupportedException(
-                               "ScriptBlocks not supported"),
-                           scriptBlockExpressionAst);
-
-                return AstVisitAction.SkipChildren;
-            }
-            */
-
-            public override AstVisitAction VisitUsingExpression(UsingExpressionAst usingExpressionAst)
-            {
-                // Using expressions are not expected in Implicit remoting commands.
-                ThrowError(new ImplicitRemotingBatchingNotSupportedException(
-                    ConsoleHostStrings.ImplicitBatchUsingNotSupported), 
-                    usingExpressionAst);
-
-                return AstVisitAction.SkipChildren;
-            }
-
-            internal static void ThrowError(ImplicitRemotingBatchingNotSupportedException ex, Ast ast)
-            {
-                InterpreterError.UpdateExceptionErrorRecordPosition(ex, ast.Extent);
-                throw ex;
-            }
-        }
-
-        internal class ImplicitRemotingBatchingNotSupportedException : RuntimeException
-        {
-            internal ImplicitRemotingBatchingNotSupportedException(string message) : base(message)
-            { }
-        }
-
-        // Commands allowed to run on target remote session along with implicit remote commands
-        private static readonly HashSet<string> AllowedCommands = new HashSet<string>(StringComparer.OrdinalIgnoreCase)
-        {
-            "ForEach-Object",
-            "Measure-Command",
-            "Measure-Object",
-            "Sort-Object",
-            "Where-Object"
-        };
-
-        // Determines if the typed command invokes implicit remoting module proxy functions in such
-        // a way as to allow simple batching, to reduce round trips between client and server sessions.
-        // Requirements:
-        //  a. All commands must be implicit remoting module proxy commands targeted to the same remote session
-        //  b. Except for *allowed* commands that can be safely run on remote session rather than client session
-        //  c. Commands must be in a simple pipeline
-        private bool TryRunAsImplicitBatch(string command, out Exception exceptionThrown)
-        {
-            exceptionThrown = null;
-
-            // Early out
-            if (_parent.RunspaceRef.IsRunspaceOverridden ||
-                (_parent.RunspaceRef.Runspace.ExecutionContext.Modules != null && !_parent.RunspaceRef.Runspace.ExecutionContext.Modules.IsImplicitRemotingModuleLoaded))
-            {
-                return false;
-            }
-
-            try
-            {
-                var scriptBlock = ScriptBlock.Create(command);
-                var scriptBlockAst = scriptBlock.Ast as ScriptBlockAst;
-                if (scriptBlockAst == null)
-                {
-                    throw new ImplicitRemotingBatchingNotSupportedException(ConsoleHostStrings.ImplicitBatchInvalidScriptBlock);
-                }
-
-                // Make sure that this is a simple pipeline
-                string errorId;
-                string errorMsg;
-                scriptBlockAst.GetSimplePipeline(true, out errorId, out errorMsg);
-                if (errorId != null)
-                {
-                    throw new ImplicitRemotingBatchingNotSupportedException(ConsoleHostStrings.ImplicitBatchScriptNotSimplePipeline);
-                }
-
-                // Run checker
-                var checker = new PipelineForBatchingChecker { ScriptBeingConverted = scriptBlockAst };
-                scriptBlockAst.InternalVisit(checker);
-
-                // If this is just a single command, there is no point in batching it
-                if (checker.Commands.Count < 2)
-                {
-                    throw new ImplicitRemotingBatchingNotSupportedException(ConsoleHostStrings.ImplicitBatchSingleCommand);
-                }
-
-                // We have a valid batching candidate
-                using (var ps = System.Management.Automation.PowerShell.Create())
-                {
-                    ps.Runspace = _parent.RunspaceRef.Runspace;
-
-                    // Check commands
-                    ps.Commands.AddCommand("Get-Command").AddParameter("Name", checker.Commands.ToArray());
-                    var cmdInfoList = ps.Invoke<CommandInfo>();
-                    if (ps.Streams.Error.Count > 0)
-                    {
-                        // Cannot continue
-                        throw new PSInvalidOperationException();
-                    }
-                    // All command modules must be implicit remoting modules from the same PSSession
-                    var success = true;
-                    var psSessionId = Guid.Empty;
-                    foreach (var cmdInfo in cmdInfoList)
-                    {
-                        // Check for allowed command
-                        var aliasInfo = cmdInfo as AliasInfo;
-                        string cmdName = (aliasInfo != null) ? aliasInfo.ReferencedCommand.Name : cmdInfo.Name;
-                        if (AllowedCommands.Contains(cmdName))
-                        {
-                            continue;
-                        }
-
-                        if (cmdInfo.Module == null || string.IsNullOrEmpty(cmdInfo.ModuleName))
-                        {
-                            success = false;
-                            break;
-                        }
-
-                        var privateData = cmdInfo.Module.PrivateData as System.Collections.Hashtable;
-                        if (privateData == null)
-                        {
-                            success = false;
-                            break;
-                        }
-                        if (!privateData.ContainsKey("ImplicitSessionId"))
-                        {
-                            success = false;
-                            break;
-                        }
-                        if (psSessionId == Guid.Empty)
-                        {
-                            psSessionId = new Guid(privateData["ImplicitSessionId"] as string);
-                        }
-                        else if (!psSessionId.ToString().Equals(privateData["ImplicitSessionId"] as string, StringComparison.OrdinalIgnoreCase))
-                        {
-                            success = false;
-                            break;
-                        }
-                    }
-
-                    if (success)
-                    {
-                        //
-                        // Invoke command pipeline as entire pipeline on remote session
-                        //
-
-                        // Update script to declare variables via Using keyword
-                        if (checker.ValidVariables.Count > 0)
-                        {
-                            foreach (var variableName in checker.ValidVariables)
-                            {
-                                command = command.Replace(variableName, ("Using:" + variableName), StringComparison.OrdinalIgnoreCase);
-                            }
-
-                            scriptBlock = ScriptBlock.Create(command);
-                        }
-
-                        // Retrieve the PSSession runspace in which to run the batch script on
-                        ps.Commands.Clear();
-                        ps.Commands.AddCommand("Get-PSSession").AddParameter("InstanceId", psSessionId);
-                        var psSession = ps.Invoke<System.Management.Automation.Runspaces.PSSession>().FirstOrDefault();
-                        if (psSession == null || ps.Streams.Error.Count > 0)
-                        {
-                            // Cannot continue
-                            throw new PSInvalidOperationException();
-                        }
-
-                        // Create and invoke implicit remoting command pipeline
-                        ps.Commands.Clear();
-                        ps.AddCommand("Invoke-Command").AddParameter("Session", psSession).AddParameter("ScriptBlock", scriptBlock).AddParameter("HideComputerName", true)
-                            .AddCommand("Out-Default");
-
-                        try
-                        {
-                            ps.Invoke();
-                        }
-                        catch (Exception ex)
-                        {
-                            var errorRecord = new ErrorRecord(ex, "ImplicitRemotingBatchExecutionTerminatingError", ErrorCategory.InvalidOperation, null);
-
-                            ps.Commands.Clear();
-                            ps.AddCommand("Write-Error").AddParameter("InputObject", errorRecord).Invoke();
-                        }
-
-                        return true;
-                    }
-                }
-            }
-            catch (Exception) { }
-
-            return false;
-        }
-
-        #endregion
 
         private Command GetOutDefaultCommand(bool endOfStatement)
         {
