+++ src/System.Management.Automation/engine/Utils.cs
@@ -6,6 +6,8 @@ using System.Runtime.InteropServices;
 using System.Diagnostics.CodeAnalysis;
 using System.Management.Automation.Configuration;
 using System.Management.Automation.Internal;
+using System.Management.Automation.Language;
+using System.Management.Automation.Runspaces;
 using System.Management.Automation.Security;
 using System.Reflection;
 using Microsoft.PowerShell.Commands;
@@ -1359,7 +1361,338 @@ namespace System.Management.Automation
             return obj != null && Marshal.IsComObject(obj);
 #endif
         }
+
+        #region Implicit Remoting Batching
+
+        // Commands allowed to run on target remote session along with implicit remote commands
+        private static readonly HashSet<string> AllowedCommands = new HashSet<string>(StringComparer.OrdinalIgnoreCase)
+        {
+            "ForEach-Object",
+            "Measure-Command",
+            "Measure-Object",
+            "Sort-Object",
+            "Where-Object"
+        };
+
+        // Determines if the typed command invokes implicit remoting module proxy functions in such
+        // a way as to allow simple batching, to reduce round trips between client and server sessions.
+        // Requirements:
+        //  a. All commands must be implicit remoting module proxy commands targeted to the same remote session
+        //  b. Except for *allowed* commands that can be safely run on remote session rather than client session
+        //  c. Commands must be in a simple pipeline
+        internal static bool TryRunAsImplicitBatch(string command, Runspace runspace)
+        {
+            try
+            {
+                var scriptBlock = ScriptBlock.Create(command);
+                var scriptBlockAst = scriptBlock.Ast as ScriptBlockAst;
+                if (scriptBlockAst == null)
+                {
+                    return false;
+                }
+
+                // Make sure that this is a simple pipeline
+                string errorId;
+                string errorMsg;
+                scriptBlockAst.GetSimplePipeline(true, out errorId, out errorMsg);
+                if (errorId != null)
+                {
+                    return false;
+                }
+
+                // Run checker
+                var checker = new PipelineForBatchingChecker { ScriptBeingConverted = scriptBlockAst };
+                scriptBlockAst.InternalVisit(checker);
+
+                // If this is just a single command, there is no point in batching it
+                if (checker.Commands.Count < 2)
+                {
+                    return false;
+                }
+
+                // We have a valid batching candidate
+                using (var ps = System.Management.Automation.PowerShell.Create())
+                {
+                    ps.Runspace = runspace;
+
+                    // Check commands
+                    ps.Commands.AddCommand("Get-Command").AddParameter("Name", checker.Commands.ToArray());
+                    var cmdInfoList = ps.Invoke<CommandInfo>();
+                    if (ps.Streams.Error.Count > 0)
+                    {
+                        // Cannot continue
+                        return false;
+                    }
+                    // All command modules must be implicit remoting modules from the same PSSession
+                    var success = true;
+                    var psSessionId = Guid.Empty;
+                    foreach (var cmdInfo in cmdInfoList)
+                    {
+                        // Check for allowed command
+                        string cmdName = (cmdInfo is AliasInfo aliasInfo) ? aliasInfo.ReferencedCommand.Name : cmdInfo.Name;
+                        if (AllowedCommands.Contains(cmdName))
+                        {
+                            continue;
+                        }
+
+                        // Commands must be from implicit remoting module
+                        if (cmdInfo.Module == null || string.IsNullOrEmpty(cmdInfo.ModuleName))
+                        {
+                            success = false;
+                            break;
+                        }
+
+                        // Commands must be from modules imported into the same remote session
+                        if (cmdInfo.Module.PrivateData is System.Collections.Hashtable privateData)
+                        {
+                            var sessionIdString = privateData["ImplicitSessionId"] as string;
+                            if (string.IsNullOrEmpty(sessionIdString))
+                            {
+                                success = false;
+                                break;
+                            }
+
+                            var sessionId = new Guid(sessionIdString);
+                            if (psSessionId == Guid.Empty)
+                            {
+                                psSessionId = sessionId;
+                            }
+                            else if (psSessionId != sessionId)
+                            {
+                                success = false;
+                                break;
+                            }
+                        }
+                        else
+                        {
+                            success = false;
+                            break;
+                        }
+                    }
+
+                    if (success)
+                    {
+                        //
+                        // Invoke command pipeline as entire pipeline on remote session
+                        //
+
+                        // Update script to declare variables via Using keyword
+                        if (checker.ValidVariables.Count > 0)
+                        {
+                            foreach (var variableName in checker.ValidVariables)
+                            {
+                                command = command.Replace(variableName, ("Using:" + variableName), StringComparison.OrdinalIgnoreCase);
+                            }
+
+                            scriptBlock = ScriptBlock.Create(command);
+                        }
+
+                        // Retrieve the PSSession runspace in which to run the batch script on
+                        ps.Commands.Clear();
+                        ps.Commands.AddCommand("Get-PSSession").AddParameter("InstanceId", psSessionId);
+                        var psSession = ps.Invoke<System.Management.Automation.Runspaces.PSSession>().FirstOrDefault();
+                        if (psSession == null || (ps.Streams.Error.Count > 0) || (psSession.Availability != RunspaceAvailability.Available))
+                        {
+                            return false;
+                        }
+
+                        // Create and invoke implicit remoting command pipeline
+                        ps.Commands.Clear();
+                        ps.AddCommand("Invoke-Command").AddParameter("Session", psSession).AddParameter("ScriptBlock", scriptBlock).AddParameter("HideComputerName", true)
+                            .AddCommand("Out-Default");
+
+                        try
+                        {
+                            ps.Invoke();
+                        }
+                        catch (Exception ex)
+                        {
+                            var errorRecord = new ErrorRecord(ex, "ImplicitRemotingBatchExecutionTerminatingError", ErrorCategory.InvalidOperation, null);
+
+                            ps.Commands.Clear();
+                            ps.AddCommand("Write-Error").AddParameter("InputObject", errorRecord).Invoke();
+                        }
+
+                        return true;
+                    }
+                }
+            }
+            catch (Exception) { }
+
+            return false;
+        }
+        #endregion
     }
+
+    #region ImplicitRemotingBatching
+
+    // A visitor to walk an AST and validate that it is a candidate for implicit remoting batching.
+    // Based on ScriptBlockToPowerShellChecker.
+    internal class PipelineForBatchingChecker : AstVisitor
+    {
+        internal readonly HashSet<string> ValidVariables = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
+        internal readonly HashSet<string> Commands = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
+        internal ScriptBlockAst ScriptBeingConverted { get; set; }
+
+        public override AstVisitAction VisitVariableExpression(VariableExpressionAst variableExpressionAst)
+        {
+            if (!variableExpressionAst.VariablePath.IsAnyLocal())
+            {
+                ThrowError(
+                        new ImplicitRemotingBatchingNotSupportedException(
+                            "VariableTypeNotSupported"),
+                        variableExpressionAst);
+            }
+
+            if (variableExpressionAst.VariablePath.UnqualifiedPath != "_")
+            {
+                ValidVariables.Add(variableExpressionAst.VariablePath.UnqualifiedPath);
+            }
+
+            return AstVisitAction.Continue;
+        }
+
+        public override AstVisitAction VisitPipeline(PipelineAst pipelineAst)
+        {
+            if (pipelineAst.PipelineElements[0] is CommandExpressionAst)
+            {
+                // If the first element is a CommandExpression, this pipeline should be the value
+                // of a parameter. We want to avoid a scriptblock that contains only a pure expression.
+                // The check "pipelineAst.Parent.Parent == ScriptBeingConverted" guarantees we throw
+                // error on that kind of scriptblock.
+
+                // Disallow pure expressions at the "top" level, but allow them otherwise.
+                // We want to catch:
+                //     1 | echo
+                // But we don't want to error out on:
+                //     echo $(1)
+                // See the comment in VisitCommand on why it's safe to check Parent.Parent, we
+                // know that we have at least:
+                //     * a NamedBlockAst (the end block)
+                //     * a ScriptBlockAst (the ast we're comparing to)
+                if (pipelineAst.GetPureExpression() == null || pipelineAst.Parent.Parent == ScriptBeingConverted)
+                {
+                    ThrowError(
+                        new ImplicitRemotingBatchingNotSupportedException(
+                            "PipelineStartingWithExpressionNotSupported"),
+                        pipelineAst);
+                }
+            }
+
+            return AstVisitAction.Continue;
+        }
+
+        public override AstVisitAction VisitCommand(CommandAst commandAst)
+        {
+            if (commandAst.InvocationOperator == TokenKind.Dot)
+            {
+                ThrowError(
+                    new ImplicitRemotingBatchingNotSupportedException(
+                        "DotSourcingNotSupported"),
+                    commandAst);
+            }
+
+            /*
+            // Up front checking ensures that we have a simple script block,
+            // so we can safely assume that the parents are:
+            //     * a PipelineAst
+            //     * a NamedBlockAst (the end block)
+            //     * a ScriptBlockAst (the ast we're comparing to)
+            // If that isn't the case, the conversion isn't allowed.  It
+            // is also safe to assume that we have at least 3 parents, a script block can't be simpler.
+            if (commandAst.Parent.Parent.Parent != ScriptBeingConverted)
+            {
+                ThrowError(
+                    new ImplicitRemotingBatchingNotSupportedException(
+                        "CantConvertWithCommandInvocations not supported"),
+                    commandAst);
+            }
+            */
+
+            if (commandAst.CommandElements[0] is ScriptBlockExpressionAst)
+            {
+                ThrowError(
+                    new ImplicitRemotingBatchingNotSupportedException(
+                        "ScriptBlockInvocationNotSupported"),
+                    commandAst);
+            }
+
+            var commandName = commandAst.GetCommandName();
+            if (commandName != null)
+            {
+                Commands.Add(commandName);
+            }
+
+            return AstVisitAction.Continue;
+        }
+
+        public override AstVisitAction VisitMergingRedirection(MergingRedirectionAst redirectionAst)
+        {
+            if (redirectionAst.ToStream != RedirectionStream.Output)
+            {
+                ThrowError(
+                    new ImplicitRemotingBatchingNotSupportedException(
+                        "MergeRedirectionNotSupported"),
+                    redirectionAst);
+            }
+
+            return AstVisitAction.Continue;
+        }
+
+        public override AstVisitAction VisitFileRedirection(FileRedirectionAst redirectionAst)
+        {
+            ThrowError(
+                new ImplicitRemotingBatchingNotSupportedException(
+                    "FileRedirectionNotSupported"),
+                redirectionAst);
+
+            return AstVisitAction.Continue;
+        }
+
+        /*
+        public override AstVisitAction VisitScriptBlockExpression(ScriptBlockExpressionAst scriptBlockExpressionAst)
+        {
+            ThrowError(new ImplicitRemotingBatchingNotSupportedException(
+                           "ScriptBlocks not supported"),
+                       scriptBlockExpressionAst);
+
+            return AstVisitAction.SkipChildren;
+        }
+        */
+
+        public override AstVisitAction VisitUsingExpression(UsingExpressionAst usingExpressionAst)
+        {
+            // Using expressions are not expected in Implicit remoting commands.
+            ThrowError(new ImplicitRemotingBatchingNotSupportedException(
+                "UsingExpressionNotSupported"),
+                usingExpressionAst);
+
+            return AstVisitAction.SkipChildren;
+        }
+
+        internal static void ThrowError(ImplicitRemotingBatchingNotSupportedException ex, Ast ast)
+        {
+            InterpreterError.UpdateExceptionErrorRecordPosition(ex, ast.Extent);
+            throw ex;
+        }
+    }
+
+    internal class ImplicitRemotingBatchingNotSupportedException : Exception
+    {
+        internal string ErrorId
+        {
+            get;
+            private set;
+        }
+
+        internal ImplicitRemotingBatchingNotSupportedException(string errorId) : base(
+            ParserStrings.ImplicitRemotingPipelineBatchingNotSupported)
+        {
+            ErrorId = errorId;
+        }
+    }
+
+    #endregion
 }
 
 namespace System.Management.Automation.Internal
@@ -1403,6 +1736,17 @@ namespace System.Management.Automation.Internal
             {
                 fieldInfo.SetValue(null, value);
             }
+        }
+
+        /// <summary>
+        /// TestImplicitRemotingBatching
+        /// </summary>
+        /// <param name="commandPipeline">Command pipeline to test</param>
+        /// <param name="runspace">Runspace to test on</param>
+        /// <returns>True if commandPipeline is batched successfully</returns>
+        public static bool TestImplicitRemotingBatching(string commandPipeline, System.Management.Automation.Runspaces.Runspace runspace)
+        {
+            return Utils.TryRunAsImplicitBatch(commandPipeline, runspace);
         }
     }
